add custom errors tracing for explorer

+++ app/data/explorer.py
@@ -1,5 +1,6 @@
-from init import curs
+from init import curs, IntegrityError
 from app.model.explorer import Explorer
+from app.errors import Missing, Duplicate
 
 
 curs.execute("""create table if not exists explorer(
@@ -9,18 +10,23 @@ curs.execute("""create table if not exists explorer(
 
 
 def row_to_model(row: tuple) -> Explorer:
-    return Explorer(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    return Explorer(name=name, country=country, description=description)
 
 
 def model_to_dict(explorer: Explorer) -> dict:
-    return explorer.dict() if explorer else None
+    return explorer.dict()
 
 
 def get_one(name: str) -> Explorer:
     query = 'select * from explorer where name=%(name)s'
     params = {'name': name}
     curs.execute(query, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f'Explorer {name} not found')
 
 
 def get_all(name: str) -> list[Explorer]:
@@ -30,27 +36,40 @@ def get_all(name: str) -> list[Explorer]:
 
 
 def create(explorer: Explorer):
+    if not explorer:
+        return None
     query = 'insert into explorer (name, country, description) \
         values (%(name)s, %(country)s, %(description)s)'
     params = model_to_dict(explorer)
-    curs.execute(query, params)
+    try:
+        curs.execute(query, params)
+    except IntegrityError:
+        raise Duplicate(msg=f'Explorer {explorer.name} already exists')
     return get_one(explorer.name)
 
 
-def modify(explorer: Explorer) -> Explorer:
+def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer):
+        return None
     query = """update explorer
-                set country=%(country)s,
-                    name=%(name)s,
-                    description=%(description)s
-                where name=%(name_orig)s"""
+            set country=%(country)s,
+                name=%(name)s,
+                description=%(description)s
+            where name=%(name_orig)s"""
     params = model_to_dict(explorer)
     params['name_orig'] = explorer.name
-    _ = curs.execute(query, params)
-    return get_one(explorer.name)
+    curs.execute(query, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f'Explorer {name} not found')
 
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name:
+        return False
     query = 'delete from explorer where name = %(name)s'
-    params = {'name': explorer.name}
-    result = curs.execute(query, params)
-    return bool(result)
+    params = {'name': name}
+    curs.execute(query, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f'Explorer {name} not found')
